python/boilerplates/django_boilerplate/requirements/base.txt
# ──────────────────────────────────────────────────────────────────────────────
# Base requirements — installed in ALL environments
# ──────────────────────────────────────────────────────────────────────────────

# Core framework
Django>=5.0,<6.0

# Environment / config
django-environ>=0.11

# REST API
djangorestframework>=3.15
django-cors-headers>=4.3

# Static files (production serving via WhiteNoise)
whitenoise>=6.7

# Image handling (for avatar uploads)
Pillow>=10.3

# PostgreSQL adapter
psycopg2-binary>=2.9

# Async task queue
celery>=5.4
redis>=5.0
django-celery-beat>=2.6      # periodic tasks / cron

# Storage (S3 / DigitalOcean / GCS)
django-storages>=1.14
boto3>=1.34                  # for AWS S3

# HTTP client
requests>=2.32

# Timezone data
tzdata>=2024.1
